Refactor BlockProcessors to use AbstractBlockProcessor
Created AbstractBlockProcessor and made rewardCoinbase method abstract
so classes that extend this can code rewardCoinbase for their needs.
